Move plugin resolvers to factories
